+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # PunyBlog changelog
 
-  - v1.8.1 (2023-11-24)
+  - v1.9.0 (2023-11-24)
+    - Feature: Adds template variables for current file being rendered
+      - `{{ render.current.basename }}` - the current base HTML file (e.g., index.html)
+      - `{{ render.current.path }}` - the relative path to the current base HTML file (e.g., path/to)
     - Maintenance: Updates README to include API documentation
     - Maintenance: Adds test cases for `template_vars` config option